修正

+++ bmi_app.py
@@ -2,7 +2,7 @@ import calc_bmi
 
 print("BMI計算アプリ")
 kg = int(input("体重："))
-h = int(input("身長："))
+h = float(input("身長："))
 print(f"BMI:{calc_bmi.get_bmi(kg,h)}")
 print(f"BMI:{calc_bmi.get_bmi(kg,h)}")
